Fixed gradient issue, and changed trips to float to get division right.

+++ VexDrawQuartz/ddw/VexDrawBinaryReader.h
@@ -10,7 +10,7 @@
 #import "VexObject.h"
 #import "Symbol.h"
 
-extern int const TWIPS;
+extern float const TWIPS;
 
 @interface VexDrawBinaryReader : NSObject
